Thêm kiểm tra nguyên liệu

+++ src/model/MonAnNguyenLieu.java
@@ -4,11 +4,11 @@ public class MonAnNguyenLieu {
     private int id;
     private int monAnId;
     private int nguyenLieuId;
-    private int soLuongCan;
+    private double soLuongCan;
 
     public MonAnNguyenLieu() {}
 
-    public MonAnNguyenLieu(int id, int monAnId, int nguyenLieuId, int soLuongCan) {
+    public MonAnNguyenLieu(int id, int monAnId, int nguyenLieuId, double soLuongCan) {
         this.id = id;
         this.monAnId = monAnId;
         this.nguyenLieuId = nguyenLieuId;
@@ -39,11 +39,11 @@ public class MonAnNguyenLieu {
         this.nguyenLieuId = nguyenLieuId;
     }
 
-    public int getSoLuongCan() {
+    public double getSoLuongCan() {
         return soLuongCan;
     }
 
-    public void setSoLuongCan(int soLuongCan) {
+    public void setSoLuongCan(double soLuongCan) {
         this.soLuongCan = soLuongCan;
     }
 // Getter, Setter
